+++ Chapter_04_Java/javafx-timeline/src/main/java/be/webtechie/timeline/DataSet.java
@@ -50,7 +50,8 @@ public enum DataSet {
             entry(2017, "JavaFX 9 controls and css modularized\n "),
             entry(2018, "JavaFX 11\n "),
             entry(2019, "JavaFX 12 and JavaFX 13\n "),
-            entry(2020, "JavaFX 14 and JavaFX 15\n "))),
+            entry(2020, "JavaFX 14 and JavaFX 15\n "),
+            entry(2021, "JavaFX 16 and JavaFX 17\n "))),
     RASPBERRYPI_BOARDS(2005, 2024, getPiBoards());
 
     private final int startYear;