+++ Chapter_06_Maven/java-maven-logging/src/main/java/be/webtechie/App.java
@@ -1,20 +1,21 @@
 package be.webtechie;
 
-import org.apache.log4j.ConsoleAppender;
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
-import org.apache.log4j.PatternLayout;
-import org.apache.log4j.RollingFileAppender;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.core.appender.ConsoleAppender;
+import org.apache.logging.log4j.core.appender.RollingFileAppender;
+import org.apache.logging.log4j.core.layout.PatternLayout;
 
 /**
  * Hello world!
  *
  */
 public class App {
-    private static Logger logger = Logger.getLogger(App.class);
+    private static Logger logger = LogManager.getLogger(App.class);
 
     public static void main( String[] args ) {
-        Logger.getRootLogger().getLoggerRepository().resetConfiguration();
+        //Logger.getRootLogger().getLoggerRepository().resetConfiguration();
         initLog();
 
         logger.debug("Debug message");
@@ -24,20 +25,24 @@ public class App {
     }
 
     private static void initLog() {
-        PatternLayout logPattern = new PatternLayout("%d{yyyyMMdd HH:mm:ss,SSS} | %-5p | [%c{1}] | %m%n");
+        PatternLayout logPattern = PatternLayout.newBuilder()
+                .withPattern("%d{yyyyMMdd HH:mm:ss,SSS} | %-5p | [%c{1}] | %m%n")
+                .build();
 
         // Log to the console, similar to System.out
-        ConsoleAppender console = new ConsoleAppender();
-        console.setName("ConsoleLogger");
-        console.setLayout(logPattern);
-        console.setThreshold(Level.DEBUG);
-        console.activateOptions();
+        ConsoleAppender console = ConsoleAppender.newBuilder()
+                .setName("ConsoleLogger")
+                .setLayout(logPattern)
+                .build();
+        logger.get
         Logger.getRootLogger().addAppender(console);
 
         // Log to a file to store it for later reference, creating max 5 files of 10MB
-        RollingFileAppender file = new RollingFileAppender();
-        file.setName("FileLogger");
-        file.setFile("logs/app.log");
+        RollingFileAppender file = RollingFileAppender.newBuilder()
+                .setName("FileLogger")
+                .setConfiguration(new )
+                .setFile("logs/app.log")
+                .setl
         file.setLayout(logPattern);
         file.setThreshold(Level.INFO);
         file.setAppend(true);