+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 This document describes all the changes in this sources-project since the book "Getting Started with Java on Rasperry Pi" was published.
 
+## 20211214 
+
+* Bumped pi4j-version to 1.3 to add support for Raspberry Pi 4/400/CM4
+* Bumped log4j-version to 2.16.0 because of security fix
+* 
 ## 20200615
 
 * Added [FOOTNOTE_LINKS.md](FOOTNOTE_LINKS.md) with all the links which can be found in the footnotes of the book.